feat: terminal launch commands can access the terminal name using '%'
Now when you specify the launch field for a language you can use the
wildcard to insert the terminal's name, appended with the name of the
workspace. This is useful if you want to have persistent terminal
sessions, ala tmux or screen. For example: "launch": "tmux new-session
-A -s % python" would launch a new python REPL within a tmux session
using the name of the terminal.